examples: removed test prefix
