+++ src/examples/Boundary.cpp
@@ -1,4 +1,4 @@
-#include "testCommon.h"
+#include "Common.h"
 
 /************************************************************************/
 /* Drawing Commands                                                     */
@@ -8,67 +8,38 @@ void tvgDrawCmds(tvg::Canvas* canvas)
 {
     if (!canvas) return;
 
-    canvas->reserve(3);                          //reserve 3 shape nodes (optional)
+    canvas->reserve(5);             //reserve 5 shape nodes (optional)
 
-    //Prepare Round Rectangle
+    //Prepare Shape1
     auto shape1 = tvg::Shape::gen();
-    shape1->appendRect(0, 0, 400, 400, 0, 0);    //x, y, w, h, rx, ry
-
-    //LinearGradient
-    auto fill = tvg::LinearGradient::gen();
-    fill->linear(0, 0, 400, 400);
-
-    //Gradient Color Stops
-    tvg::Fill::ColorStop colorStops[2];
-    colorStops[0] = {0, 0, 0, 0, 255};
-    colorStops[1] = {1, 255, 255, 255, 255};
-
-    fill->colorStops(colorStops, 2);
-
-    shape1->fill(move(fill));
+    shape1->appendRect(-100, -100, 1000, 1000, 50, 50);
+    shape1->fill(255, 255, 255, 255);
     if (canvas->push(move(shape1)) != tvg::Result::Success) return;
 
-    //Prepare Circle
+    //Prepare Shape2
     auto shape2 = tvg::Shape::gen();
-    shape2->appendCircle(400, 400, 200, 200);    //cx, cy, radiusW, radiusH
-
-    //LinearGradient
-    auto fill2 = tvg::LinearGradient::gen();
-    fill2->linear(400, 200, 400, 600);
-
-    //Gradient Color Stops
-    tvg::Fill::ColorStop colorStops2[3];
-    colorStops2[0] = {0, 255, 0, 0, 255};
-    colorStops2[1] = {0.5, 255, 255, 0, 255};
-    colorStops2[2] = {1, 255, 255, 255, 255};
-
-    fill2->colorStops(colorStops2, 3);
-
-    shape2->fill(move(fill2));
+    shape2->appendRect(-100, -100, 250, 250, 50, 50);
+    shape2->fill(0, 0, 255, 255);
     if (canvas->push(move(shape2)) != tvg::Result::Success) return;
 
-
-    //Prepare Ellipse
+    //Prepare Shape3
     auto shape3 = tvg::Shape::gen();
-    shape3->appendCircle(600, 600, 150, 100);    //cx, cy, radiusW, radiusH
-
-    //LinearGradient
-    auto fill3 = tvg::LinearGradient::gen();
-    fill3->linear(450, 600, 750, 600);
-
-    //Gradient Color Stops
-    tvg::Fill::ColorStop colorStops3[4];
-    colorStops3[0] = {0, 0, 127, 0, 127};
-    colorStops3[1] = {0.25, 0, 170, 170, 170};
-    colorStops3[2] = {0.5, 200, 0, 200, 200};
-    colorStops3[3] = {1, 255, 255, 255, 255};
-
-    fill3->colorStops(colorStops3, 4);
-
-    shape3->fill(move(fill3));
+    shape3->appendRect(500, 500, 550, 550, 0, 0);
+    shape3->fill(0, 255, 255, 255);
     if (canvas->push(move(shape3)) != tvg::Result::Success) return;
-}
 
+    //Prepare Shape4
+    auto shape4 = tvg::Shape::gen();
+    shape4->appendCircle(800, 100, 200, 200);
+    shape4->fill(255, 255, 0, 255);
+    if (canvas->push(move(shape4)) != tvg::Result::Success) return;
+
+    //Prepare Shape5
+    auto shape5 = tvg::Shape::gen();
+    shape5->appendCircle(200, 650, 250, 200);
+    shape5->fill(0, 0, 0, 255);
+    if (canvas->push(move(shape5)) != tvg::Result::Success) return;
+}
 
 /************************************************************************/
 /* Sw Engine Test Code                                                  */
