+++ src/examples/Common.h
