+++ src/examples/PathCopy.cpp
@@ -1,4 +1,4 @@
-#include "testCommon.h"
+#include "Common.h"
 
 /************************************************************************/
 /* Drawing Commands                                                     */
@@ -8,37 +8,80 @@ void tvgDrawCmds(tvg::Canvas* canvas)
 {
     if (!canvas) return;
 
-    canvas->reserve(5);             //reserve 5 shape nodes (optional)
+    /* Star */
 
-    //Prepare Shape1
+    //Prepare Path Commands
+    tvg::PathCommand cmds[11];
+    cmds[0] = tvg::PathCommand::MoveTo;
+    cmds[1] = tvg::PathCommand::LineTo;
+    cmds[2] = tvg::PathCommand::LineTo;
+    cmds[3] = tvg::PathCommand::LineTo;
+    cmds[4] = tvg::PathCommand::LineTo;
+    cmds[5] = tvg::PathCommand::LineTo;
+    cmds[6] = tvg::PathCommand::LineTo;
+    cmds[7] = tvg::PathCommand::LineTo;
+    cmds[8] = tvg::PathCommand::LineTo;
+    cmds[9] = tvg::PathCommand::LineTo;
+    cmds[10] = tvg::PathCommand::Close;
+
+    //Prepare Path Points
+    tvg::Point pts[10];
+    pts[0] = {199, 34};    //MoveTo
+    pts[1] = {253, 143};   //LineTo
+    pts[2] = {374, 160};   //LineTo
+    pts[3] = {287, 244};   //LineTo
+    pts[4] = {307, 365};   //LineTo
+    pts[5] = {199, 309};   //LineTo
+    pts[6] = {97, 365};    //LineTo
+    pts[7] = {112, 245};   //LineTo
+    pts[8] = {26, 161};    //LineTo
+    pts[9] = {146, 143};   //LineTo
+
     auto shape1 = tvg::Shape::gen();
-    shape1->appendRect(-100, -100, 1000, 1000, 50, 50);
-    shape1->fill(255, 255, 255, 255);
+    shape1->appendPath(cmds, 11, pts, 10);     //copy path data
+    shape1->fill(0, 255, 0, 255);
     if (canvas->push(move(shape1)) != tvg::Result::Success) return;
 
-    //Prepare Shape2
+    /* Circle */
+    auto cx = 550.0f;
+    auto cy = 550.0f;
+    auto radius = 125.0f;
+    auto halfRadius = radius * 0.552284f;
+
+    //Prepare Path Commands
+    tvg::PathCommand cmds2[6];
+    cmds2[0] = tvg::PathCommand::MoveTo;
+    cmds2[1] = tvg::PathCommand::CubicTo;
+    cmds2[2] = tvg::PathCommand::CubicTo;
+    cmds2[3] = tvg::PathCommand::CubicTo;
+    cmds2[4] = tvg::PathCommand::CubicTo;
+    cmds2[5] = tvg::PathCommand::Close;
+
+    //Prepare Path Points
+    tvg::Point pts2[13];
+    pts2[0] = {cx, cy - radius};    //MoveTo
+    //CubicTo 1
+    pts2[1] = {cx + halfRadius, cy - radius};      //Ctrl1
+    pts2[2] = {cx + radius, cy - halfRadius};      //Ctrl2
+    pts2[3] = {cx + radius, cy};                   //To
+    //CubicTo 2
+    pts2[4] = {cx + radius, cy + halfRadius};      //Ctrl1
+    pts2[5] = {cx + halfRadius, cy + radius};      //Ctrl2
+    pts2[6] = {cx, cy+ radius};                    //To
+    //CubicTo 3
+    pts2[7] = {cx - halfRadius, cy + radius};      //Ctrl1
+    pts2[8] = {cx - radius, cy + halfRadius};      //Ctrl2
+    pts2[9] = {cx - radius, cy};                   //To
+    //CubicTo 4
+    pts2[10] = {cx - radius, cy - halfRadius};     //Ctrl1
+    pts2[11] = {cx - halfRadius, cy - radius};     //Ctrl2
+    pts2[12] = {cx, cy - radius};                  //To
+
     auto shape2 = tvg::Shape::gen();
-    shape2->appendRect(-100, -100, 250, 250, 50, 50);
-    shape2->fill(0, 0, 255, 255);
+    shape2->appendPath(cmds2, 6, pts2, 13);     //copy path data
+    shape2->fill(255, 255, 0, 255);
     if (canvas->push(move(shape2)) != tvg::Result::Success) return;
 
-    //Prepare Shape3
-    auto shape3 = tvg::Shape::gen();
-    shape3->appendRect(500, 500, 550, 550, 0, 0);
-    shape3->fill(0, 255, 255, 255);
-    if (canvas->push(move(shape3)) != tvg::Result::Success) return;
-
-    //Prepare Shape4
-    auto shape4 = tvg::Shape::gen();
-    shape4->appendCircle(800, 100, 200, 200);
-    shape4->fill(255, 255, 0, 255);
-    if (canvas->push(move(shape4)) != tvg::Result::Success) return;
-
-    //Prepare Shape5
-    auto shape5 = tvg::Shape::gen();
-    shape5->appendCircle(200, 650, 250, 200);
-    shape5->fill(0, 0, 0, 255);
-    if (canvas->push(move(shape5)) != tvg::Result::Success) return;
 }
 
 /************************************************************************/
