+++ src/examples/SceneTransform.cpp
@@ -1,56 +1,108 @@
-#include "testCommon.h"
+#include "Common.h"
 
 /************************************************************************/
 /* Drawing Commands                                                     */
 /************************************************************************/
-tvg::Shape* pShape = nullptr;
+tvg::Scene* pScene1 = nullptr;
+tvg::Scene* pScene2 = nullptr;
 
 void tvgDrawCmds(tvg::Canvas* canvas)
 {
     if (!canvas) return;
 
-    //Shape (for BG)
-    auto bg = tvg::Shape::gen();
-    bg->appendRect(0, 0, WIDTH, HEIGHT, 0, 0);
+    //Create a Scene1
+    auto scene = tvg::Scene::gen();
+    pScene1 = scene.get();
+    scene->reserve(3);   //reserve 3 shape nodes (optional)
 
-    //fill property will be retained
-    bg->fill(255, 255, 255, 255);
+    //Prepare Round Rectangle (Scene1)
+    auto shape1 = tvg::Shape::gen();
+    shape1->appendRect(-235, -250, 400, 400, 50, 50);  //x, y, w, h, rx, ry
+    shape1->fill(0, 255, 0, 255);                      //r, g, b, a
+    shape1->stroke(5);                                 //width
+    shape1->stroke(255, 255, 255, 255);                //r, g, b, a
+    scene->push(move(shape1));
 
-    if (canvas->push(move(bg)) != tvg::Result::Success) return;
+    //Prepare Circle (Scene1)
+    auto shape2 = tvg::Shape::gen();
+    shape2->appendCircle(-165, -150, 200, 200);    //cx, cy, radiusW, radiusH
+    shape2->fill(255, 255, 0, 255);                //r, g, b, a
+    scene->push(move(shape2));
 
-    //Shape
-    auto shape = tvg::Shape::gen();
+    //Prepare Ellipse (Scene1)
+    auto shape3 = tvg::Shape::gen();
+    shape3->appendCircle(265, 250, 150, 100);      //cx, cy, radiusW, radiusH
+    shape3->fill(0, 255, 255, 255);                //r, g, b, a
+    scene->push(move(shape3));
 
-    /* Acquire shape pointer to access it again.
-       instead, you should consider not to interrupt this pointer life-cycle. */
-    pShape = shape.get();
+    scene->translate(350, 350);
+    scene->scale(0.5);
 
-    shape->appendRect(-100, -100, 200, 200, 0, 0);
+    //Create Scene2
+    auto scene2 = tvg::Scene::gen();
+    pScene2 = scene2.get();
+    scene2->reserve(2);   //reserve 2 shape nodes (optional)
 
-    //fill property will be retained
-    shape->fill(127, 255, 255, 255);
-    shape->stroke(0, 0, 255, 255);
-    shape->stroke(1);
+    //Star (Scene2)
+    auto shape4 = tvg::Shape::gen();
 
-    if (canvas->push(move(shape)) != tvg::Result::Success) return;
+    //Appends Paths
+    shape4->moveTo(0, -114.5);
+    shape4->lineTo(54, -5.5);
+    shape4->lineTo(175, 11.5);
+    shape4->lineTo(88, 95.5);
+    shape4->lineTo(108, 216.5);
+    shape4->lineTo(0, 160.5);
+    shape4->lineTo(-102, 216.5);
+    shape4->lineTo(-87, 96.5);
+    shape4->lineTo(-173, 12.5);
+    shape4->lineTo(-53, -5.5);
+    shape4->close();
+    shape4->fill(0, 0, 255, 127);
+    shape4->stroke(3);                             //width
+    shape4->stroke(0, 0, 255, 255);                //r, g, b, a
+    scene2->push(move(shape4));
+
+    //Circle (Scene2)
+    auto shape5 = tvg::Shape::gen();
+
+    auto cx = -150.0f;
+    auto cy = -150.0f;
+    auto radius = 100.0f;
+    auto halfRadius = radius * 0.552284f;
+
+    //Append Paths
+    shape5->moveTo(cx, cy - radius);
+    shape5->cubicTo(cx + halfRadius, cy - radius, cx + radius, cy - halfRadius, cx + radius, cy);
+    shape5->cubicTo(cx + radius, cy + halfRadius, cx + halfRadius, cy + radius, cx, cy+ radius);
+    shape5->cubicTo(cx - halfRadius, cy + radius, cx - radius, cy + halfRadius, cx - radius, cy);
+    shape5->cubicTo(cx - radius, cy - halfRadius, cx - halfRadius, cy - radius, cx, cy - radius);
+    shape5->close();
+    shape5->fill(255, 0, 0, 127);
+    scene2->push(move(shape5));
+
+    scene2->translate(500, 350);
+
+    //Push scene2 onto the scene
+    scene->push(move(scene2));
+
+    //Draw the Scene onto the Canvas
+    canvas->push(move(scene));
 }
 
 void tvgUpdateCmds(tvg::Canvas* canvas, float progress)
 {
     if (!canvas) return;
 
-    /* Update shape directly.
-       You can update only necessary properties of this shape,
+    /* Update scene directly.
+       You can update only necessary properties of this scene,
        while retaining other properties. */
 
-    //Reset Shape
-    if (pShape->reset() == tvg::Result::Success) {
-        pShape->appendRect(-100 + (800 * progress), -100 + (800 * progress), 200, 200, (100 * progress), (100 * progress));
-        pShape->stroke(30 * progress);
+    pScene1->rotate(360 * progress);
+    pScene2->rotate(360 * progress);
 
-        //Update shape for drawing (this may work asynchronously)
-        canvas->update(pShape);
-    }
+    //Update shape for drawing (this may work asynchronously)
+    canvas->update(pScene1);
 }
