+++ src/examples/Stress.cpp
@@ -1,54 +1,80 @@
-#include "testCommon.h"
+#include <vector>
+#include "Common.h"
 
 /************************************************************************/
 /* Drawing Commands                                                     */
 /************************************************************************/
 
+#define NUM_PER_LINE 16
+#define SIZE 50
+
+static int count = 0;
+static int frame = 0;
+static std::vector<tvg::Picture*> pictures;
+static double t1, t2, t3, t4;
+
+void svgDirCallback(const char* name, const char* path, void* data)
+{
+    auto picture = tvg::Picture::gen();
+
+    char buf[PATH_MAX];
+    sprintf(buf, "/%s/%s", path, name);
+
+    if (picture->load(buf) != tvg::Result::Success) return;
+
+    float x, y, w, h;
+    picture->viewbox(&x, &y, &w, &h);
+
+    float rate = (SIZE/(w > h ? w : h));
+    picture->scale(rate);
+
+    x *= rate;
+    y *= rate;
+    w *= rate;
+    h *= rate;
+
+    //Center Align ?
+    if (w > h) {
+         y -= (SIZE - h) * 0.5f;
+    } else {
+         x -= (SIZE - w) * 0.5f;
+    }
+
+    picture->translate((count % NUM_PER_LINE) * SIZE - x, SIZE * (count / NUM_PER_LINE) - y);
+    ++count;
+
+    //Duplicates
+    for (int i = 0; i < NUM_PER_LINE - 1; i++) {
+        tvg::Picture* dup = static_cast<tvg::Picture*>(picture->duplicate());
+        dup->translate((count % NUM_PER_LINE) * SIZE - x, SIZE * (count / NUM_PER_LINE) - y);
+        pictures.push_back(dup);
+        ++count;
+    }
+
+    cout << "SVG: " << buf << endl;
+    pictures.push_back(picture.release());    
+}
+
 void tvgDrawCmds(tvg::Canvas* canvas)
 {
     if (!canvas) return;
 
-    canvas->reserve(5);
+    //Background
+    auto shape = tvg::Shape::gen();
+    shape->appendRect(0, 0, WIDTH, HEIGHT, 0, 0);    //x, y, w, h, rx, ry
+    shape->fill(255, 255, 255, 255);                 //r, g, b, a
 
-    //Prepare Round Rectangle
-    auto shape1 = tvg::Shape::gen();
-    shape1->appendRect(0, 0, 400, 400, 50, 50);  //x, y, w, h, rx, ry
-    shape1->fill(0, 255, 0, 255);                //r, g, b, a
-    if (canvas->push(move(shape1)) != tvg::Result::Success) return;
+    if (canvas->push(move(shape)) != tvg::Result::Success) return;
 
-    //Prepare Circle
-    auto shape2 = tvg::Shape::gen();
-    shape2->appendCircle(400, 400, 200, 200);    //cx, cy, radiusW, radiusH
-    shape2->fill(255, 255, 0, 170);              //r, g, b, a
-    if (canvas->push(move(shape2)) != tvg::Result::Success) return;
+    eina_file_dir_list(EXAMPLE_DIR, EINA_TRUE, svgDirCallback, canvas);
 
-    //Prepare Ellipse
-    auto shape3 = tvg::Shape::gen();
-    shape3->appendCircle(400, 400, 250, 100);    //cx, cy, radiusW, radiusH
-    shape3->fill(255, 255, 255, 100);            //r, g, b, a
-    if (canvas->push(move(shape3)) != tvg::Result::Success) return;
-
-    //Prepare Star
-    auto shape4 = tvg::Shape::gen();
-    shape4->moveTo(199, 234);
-    shape4->lineTo(253, 343);
-    shape4->lineTo(374, 360);
-    shape4->lineTo(287, 444);
-    shape4->lineTo(307, 565);
-    shape4->lineTo(199, 509);
-    shape4->lineTo(97, 565);
-    shape4->lineTo(112, 445);
-    shape4->lineTo(26, 361);
-    shape4->lineTo(146, 343);
-    shape4->close();
-    shape4->fill(255, 0, 200, 200);
-    if (canvas->push(move(shape4)) != tvg::Result::Success) return;
-
-    //Prepare Opaque Ellipse
-    auto shape5 = tvg::Shape::gen();
-    shape5->appendCircle(600, 650, 200, 150);
-    shape5->fill(0, 0, 255, 255);
-    if (canvas->push(move(shape5)) != tvg::Result::Success) return;
+    /* This showcase shows you asynchrounous loading of svg.
+       For this, pushing pictures at a certian sync time.
+       This means it earns the time to finish loading svg resources,
+       otherwise you can push pictures immediately. */
+    for (auto picture : pictures) {
+        canvas->push(unique_ptr<tvg::Picture>(picture));
+    }
 }
 
 
@@ -73,9 +99,32 @@ void tvgSwTest(uint32_t* buffer)
 
 void drawSwView(void* data, Eo* obj)
 {
+    t3 = ecore_time_get();
+
+    //Drawing task can be performed asynchronously.
     if (swCanvas->draw() == tvg::Result::Success) {
         swCanvas->sync();
     }
+
+    t4 = ecore_time_get();
+    printf("[%5d]: total[%fs] update[%fs], render[%fs]\n", ++frame, t4 - t1, t2 - t1, t4 - t3);
+}
+
+void transitSwCb(Elm_Transit_Effect *effect, Elm_Transit* transit, double progress)
+{
+    t1 = ecore_time_get();
+
+    for (auto picture : pictures) {
+        picture->rotate(progress * 360);
+        swCanvas->update(picture);
+    }
+
+    t2 = ecore_time_get();
+
+    //Update Efl Canvas
+    auto img = (Eo*) effect;
+    evas_object_image_pixels_dirty_set(img, EINA_TRUE);
+    evas_object_image_data_update_add(img, 0, 0, WIDTH, HEIGHT);    
 }
 
 
@@ -111,6 +160,14 @@ void drawGLview(Evas_Object *obj)
     }
 }
 
+void transitGlCb(Elm_Transit_Effect *effect, Elm_Transit* transit, double progress)
+{
+    for (auto picture : pictures) {
+        picture->rotate(progress * 360);
+        glCanvas->update(picture);
+    }
+}
+
 
 /************************************************************************/
 /* Main Code                                                            */
@@ -139,17 +196,25 @@ int main(int argc, char **argv)
 
         elm_init(argc, argv);
 
+        Elm_Transit *transit = elm_transit_add();
+
         if (tvgEngine == tvg::CanvasEngine::Sw) {
-            createSwView();
+            auto view = createSwView();
+            elm_transit_effect_add(transit, transitSwCb, view, nullptr);
         } else {
-            createGlView();
+            auto view = createGlView();
+            elm_transit_effect_add(transit, transitGlCb, view, nullptr);
         }
+
+        elm_transit_duration_set(transit, 2);
+        elm_transit_repeat_times_set(transit, -1);
+        elm_transit_go(transit);
 
         elm_run();
         elm_shutdown();
 
         //Terminate ThorVG Engine
-        tvg::Initializer::term(tvgEngine);
+        tvg::Initializer::term(tvg::CanvasEngine::Sw);
 
     } else {
         cout << "engine is not supported" << endl;
