+++ src/examples/Transform.cpp
@@ -1,75 +1,75 @@
-#include <vector>
-#include "testCommon.h"
+#include "Common.h"
 
 /************************************************************************/
 /* Drawing Commands                                                     */
 /************************************************************************/
-
-#define NUM_PER_LINE 4
-#define SIZE 200
-
-static int count = 0;
-
-static std::vector<unique_ptr<tvg::Picture>> pictures;
-
-void svgDirCallback(const char* name, const char* path, void* data)
-{
-    auto picture = tvg::Picture::gen();
-
-    char buf[PATH_MAX];
-    sprintf(buf, "/%s/%s", path, name);
-
-    if (picture->load(buf) != tvg::Result::Success) return;
-
-    float x, y, w, h;
-    picture->viewbox(&x, &y, &w, &h);
-
-    float rate = (SIZE/(w > h ? w : h));
-    picture->scale(rate);
-
-    x *= rate;
-    y *= rate;
-    w *= rate;
-    h *= rate;
-
-    //Center Align ?
-    if (w > h) {
-         y -= (SIZE - h) * 0.5f;
-    } else {
-         x -= (SIZE - w) * 0.5f;
-    }
-
-    picture->translate((count % NUM_PER_LINE) * SIZE - x, SIZE * (count / NUM_PER_LINE) - y);
-
-    pictures.push_back(move(picture));
-
-    cout << "SVG: " << buf << endl;
-
-    count++;
-}
+tvg::Shape* pShape = nullptr;
+tvg::Shape* pShape2 = nullptr;
+tvg::Shape* pShape3 = nullptr;
 
 void tvgDrawCmds(tvg::Canvas* canvas)
 {
     if (!canvas) return;
 
-    //Background
+    //Shape1
     auto shape = tvg::Shape::gen();
-    shape->appendRect(0, 0, WIDTH, HEIGHT, 0, 0);    //x, y, w, h, rx, ry
-    shape->fill(255, 255, 255, 255);                 //r, g, b, a
 
+    /* Acquire shape pointer to access it again.
+       instead, you should consider not to interrupt this pointer life-cycle. */
+    pShape = shape.get();
+
+    shape->appendRect(-285, -300, 200, 200, 0, 0);
+    shape->appendRect(-185, -200, 300, 300, 100, 100);
+    shape->appendCircle(115, 100, 100, 100);
+    shape->appendCircle(115, 200, 170, 100);
+    shape->fill(255, 255, 255, 255);
+    shape->translate(385, 400);
     if (canvas->push(move(shape)) != tvg::Result::Success) return;
 
-    eina_file_dir_list(EXAMPLE_DIR, EINA_TRUE, svgDirCallback, canvas);
+    //Shape2
+    auto shape2 = tvg::Shape::gen();
+    pShape2 = shape2.get();
+    shape2->appendRect(-50, -50, 100, 100, 0, 0);
+    shape2->fill(0, 255, 255, 255);
+    shape2->translate(400, 400);
+    if (canvas->push(move(shape2)) != tvg::Result::Success) return;
 
-    /* This showcase shows you asynchrounous loading of svg.
-       For this, pushing pictures at a certian sync time.
-       This means it earns the time to finish loading svg resources,
-       otherwise you can push pictures immediately. */
-    for (auto& paint : pictures) {
-        canvas->push(move(paint));
-    }
+    //Shape3
+    auto shape3 = tvg::Shape::gen();
+    pShape3 = shape3.get();
 
-    pictures.clear();
+    /* Look, how shape3's origin is different with shape2
+       The center of the shape is the anchor point for transformation. */
+    shape3->appendRect(100, 100, 150, 50, 20, 20);
+    shape3->fill(255, 0, 255, 255);
+    shape3->translate(400, 400);
+    if (canvas->push(move(shape3)) != tvg::Result::Success) return;
+}
+
+void tvgUpdateCmds(tvg::Canvas* canvas, float progress)
+{
+    if (!canvas) return;
+
+    /* Update shape directly.
+       You can update only necessary properties of this shape,
+       while retaining other properties. */
+
+    //Update Shape1
+    pShape->scale(1 - 0.75 * progress);
+    pShape->rotate(360 * progress);
+
+    //Update shape for drawing (this may work asynchronously)
+    if (canvas->update(pShape) != tvg::Result::Success) return;
+
+    //Update Shape2
+    pShape2->rotate(360 * progress);
+    pShape2->translate(400 + progress * 300, 400);
+    if (canvas->update(pShape2) != tvg::Result::Success) return;
+
+    //Update Shape3
+    pShape3->rotate(-360 * progress);
+    pShape3->scale(0.5 + progress);
+    if (canvas->update(pShape3) != tvg::Result::Success) return;
 }
 
 
@@ -90,6 +90,16 @@ void tvgSwTest(uint32_t* buffer)
        internal data asynchronously for coming rendering.
        Canvas keeps this shape node unless user call canvas->clear() */
     tvgDrawCmds(swCanvas.get());
+}
+
+void transitSwCb(Elm_Transit_Effect *effect, Elm_Transit* transit, double progress)
+{
+    tvgUpdateCmds(swCanvas.get(), progress);
+
+    //Update Efl Canvas
+    Eo* img = (Eo*) effect;
+    evas_object_image_data_update_add(img, 0, 0, WIDTH, HEIGHT);
+    evas_object_image_pixels_dirty_set(img, EINA_TRUE);
 }
 
 void drawSwView(void* data, Eo* obj)
@@ -132,6 +142,11 @@ void drawGLview(Evas_Object *obj)
     }
 }
 
+void transitGlCb(Elm_Transit_Effect *effect, Elm_Transit* transit, double progress)
+{
+    tvgUpdateCmds(glCanvas.get(), progress);
+}
+
 
 /************************************************************************/
 /* Main Code                                                            */
@@ -160,17 +175,26 @@ int main(int argc, char **argv)
 
         elm_init(argc, argv);
 
+        Elm_Transit *transit = elm_transit_add();
+
         if (tvgEngine == tvg::CanvasEngine::Sw) {
-            createSwView();
+            auto view = createSwView();
+            elm_transit_effect_add(transit, transitSwCb, view, nullptr);
         } else {
-            createGlView();
+            auto view = createGlView();
+            elm_transit_effect_add(transit, transitGlCb, view, nullptr);
         }
+
+        elm_transit_duration_set(transit, 2);
+        elm_transit_repeat_times_set(transit, -1);
+        elm_transit_auto_reverse_set(transit, EINA_TRUE);
+        elm_transit_go(transit);
 
         elm_run();
         elm_shutdown();
 
         //Terminate ThorVG Engine
-        tvg::Initializer::term(tvg::CanvasEngine::Sw);
+        tvg::Initializer::term(tvgEngine);
 
     } else {
         cout << "engine is not supported" << endl;
