+++ src/examples/meson.build
@@ -1,32 +1,32 @@
 examples_dep = dependency('elementary', required : true)
 
 source_file = [
-    'testArc.cpp',
-    'testAsync.cpp',
-    'testBlending.cpp',
-    'testBoundary.cpp',
-    'testCapi.cpp',
-    'testCustomTransform.cpp',
-    'testDirectUpdate.cpp',
-    'testDuplicate.cpp',
-    'testGradientTransform.cpp',
-    'testLinearGradient.cpp',
-    'testMultiCanvas.cpp',
-    'testMultiShapes.cpp',
-    'testPathCopy.cpp',
-    'testPath.cpp',
-    'testRadialGradient.cpp',
-    'testScene.cpp',
-    'testSceneTransform.cpp',
-    'testShape.cpp',
-    'testStress.cpp',
-    'testStroke.cpp',
-    'testStrokeLine.cpp',
-    'testSvg2.cpp',
-    'testSvg.cpp',
-    'testTransform.cpp',
-    'testUpdate.cpp',
-    'testClipPath.cpp',
+    'Arc.cpp',
+    'Async.cpp',
+    'Blending.cpp',
+    'Boundary.cpp',
+    'Capi.cpp',
+    'CustomTransform.cpp',
+    'DirectUpdate.cpp',
+    'Duplicate.cpp',
+    'GradientTransform.cpp',
+    'LinearGradient.cpp',
+    'MultiCanvas.cpp',
+    'MultiShapes.cpp',
+    'PathCopy.cpp',
+    'Path.cpp',
+    'RadialGradient.cpp',
+    'Scene.cpp',
+    'SceneTransform.cpp',
+    'Shape.cpp',
+    'Stress.cpp',
+    'Stroke.cpp',
+    'StrokeLine.cpp',
+    'Svg2.cpp',
+    'Svg.cpp',
+    'Transform.cpp',
+    'Update.cpp',
+    'ClipPath.cpp',
 ]
 
 foreach current_file : source_file
